// src/DomainLayer/Models/Fragment/FragmentEntity.ts
export type UserInfoBasic = {
  Email: string;
  Doc: string;
  NickName: string;
  FirstName: string;
  LastName: string;
  Gender: string;
  PhoneNumber: string;
  Address: string;
  Profile: string;
  Verified: boolean;
};

export type ProfileImg = {
  ImgUser: string;
  ImgFrontPage: string;
  ImgWallpaper: string;
};

export type OneValue = {
  Value: any;
};